add LIS
emmm. i can't understand the contents of the book about LIS ...
have no choice but to debug the code..
finally, i know what the book want to say
but why i can't understand?
i think that the book always write the final result directly, not to say how to
think this problem, not to write the process of thinking.